requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Werkzeug==2.3.6
blinker==1.6.2
Flask-WTF==1.2.2
pyyaml